configured distribution-installed commands for built-in measurements
s.t. these may be scheduled as built-in tasks

for simplicity of development, built-in modules are here integrated into
the netrics library, rather than made into complete stand-alone scripts
and merely copied into place.

NOTE: lml is *not* here affected. it is undergoing development in a
separate branch, and may be incorporated upon merge.