completed last-mile latency measurement: scamper trace version
Adds both task `lml` (executable `netrics-lml`) and task alias
`lml-scamper` (`netrics-lml-scamper`) for clarity.

This version of the measurement is added to the default configuration
but commented out as it may not be enabled by default.

For speed and niceness, this lml test will randomly select an endpoint
to trace, from a default set of Google DNS as well as CloudFlare DNS.
(The former is generally speedier to respond; regardless, it might be
"nice" to round-robin, and for that matter to fall back in the event of
response failure.)

Otherwise:

Measurements' schedules in default configuration now use hashed cron
expressions to *discourage* their collision (though any which *may not*
run concurrently will be configured as such).

completes #13

part of #3