last-mile latency measurement: traceroute+ping version
Adds task `lml-traceroute` (executable `netrics-lml-traceroute`).

This task mirrors the `lml-scamper` version already published. It
differs in that the endpoint is traced with `traceroute` and then
`ping`'d; (and, traceroute output must be specially parsed, which is
relatively laborious).

As with `lml-scamper`, for speed and niceness, `lml-traceroute` will
randomly select an endpoint to trace, from a default set of Google DNS
as well as CloudFlare DNS. (The former is generally speedier to respond;
regardless, it might be "nice" to round-robin, and for that matter to
fall back in the event of response failure.)

completes #25

part of #11

part of #3