remove draft traceroute (scamper) measurement
this measurement was drafted during an experimental phase of
development, and no longer reflects planned work.

it is superceded by the `lml` and/or `hops` measurements, which make use
of traceroute/scamper for their own purposes.

this measurement does neither: in fact it recorded trace data without
derivative. this might be useful at some point, but is not an extant
component of netrics built-in measurements.

resolves #18