fix: speed-ookla shouldn't treat CWD as HOME
unset HOME bug was fixed only by inviting ookla speedtest to treat
current working directory as the user HOME. this is not desired --
especially not in a systemd environment (where this is likely the
filesystem root directory).

instead, HOME is set to a temporary directory.

bump version 0.0.1-rc.16 → 0.0.1-rc.17

see #48