speed-ndt7: added measurement to default configuration
The NDT7 test is made the model for speed test configuration (by the
alphabet and as it requires no additional paramters).

speed-ookla is intended to have the same configuration, with additional
parameters -- and so YAML anchor/alias is used.

Fate is upgraded to receive a bug-fix fully supporting this YAML
feature.

See #23